Move fixed-point-builder out of CIRGenExprScalar.cpp
